+++ src/config.h
@@ -195,6 +195,7 @@ typedef struct {
   bool short_long_opts;
   bool bold_as_special;
   int selection_show_size;
+  bool hover_title;
   // Legacy
   bool use_system_colours;
   bool old_bold;